SIMPLY-2510 Group related methods together in sign-in VCs
