+++ Simplified/SignInLogic/NYPLSignInBusinessLogic+BookmarkSyncing.swift
@@ -1,0 +1,77 @@
+//
+//  NYPLSignInBusinessLogic+BookmarkSyncing.swift
+//  Simplified
+//
+//  Created by Ettore Pasquini on 11/4/20.
+//  Copyright © 2020 NYPL Labs. All rights reserved.
+//
+
+import Foundation
+
+extension NYPLSignInBusinessLogic {
+  @objc func shouldShowSyncButton() -> Bool {
+    guard let libraryDetails = libraryAccount?.details else {
+      return false
+    }
+
+    return libraryDetails.supportsSimplyESync &&
+      libraryDetails.getLicenseURL(.annotations) != nil &&
+      userAccount.hasCredentials() &&
+      libraryAccountID == libraryAccountsProvider.currentAccount?.uuid
+  }
+
+  /// Updates server sync setting for the currently selected library.
+  /// - Parameters:
+  ///   - granted: Whether the user is granting sync permission or not.
+  ///   - postServerSyncCompletion: Only run when granting sync permission.
+  @objc func changeSyncPermission(to granted: Bool,
+                                  postServerSyncCompletion: @escaping (Bool) -> Void) {
+    if granted {
+      // When granting, attempt to enable on the server.
+      NYPLAnnotations.updateServerSyncSetting(toEnabled: true) { success in
+        self.libraryAccount?.details?.syncPermissionGranted = success
+        postServerSyncCompletion(success)
+      }
+    } else {
+      // When revoking, just ignore the server's annotations.
+      libraryAccount?.details?.syncPermissionGranted = false
+    }
+  }
+
+  /// Checks with the annotations sync status with the server, adding logic
+  /// to make sure only one such requests is being executed at a time.
+  /// - Parameters:
+  ///   - preWork: Any preparatory work to be done. This block is run
+  ///   synchronously on the main thread. It's not run at all if a request is
+  ///   already ongoing or if the current library doesn't support syncing.
+  ///   - postWork: Any final work to be done. This block is run
+  ///   on the main thread. It's not run at all if a request is
+  ///   already ongoing or if the current library doesn't support syncing.
+  @objc func checkSyncPermission(preWork: () -> Void,
+                                 postWork: @escaping (_ enableSync: Bool) -> Void) {
+    guard let libraryDetails = libraryAccount?.details else {
+      return
+    }
+
+    guard permissionsCheckLock.try(), libraryDetails.supportsSimplyESync else {
+      Log.debug(#file, "Skipping sync setting check. Request already in progress or sync not supported.")
+      return
+    }
+
+    NYPLMainThreadRun.sync {
+      preWork()
+    }
+
+    NYPLAnnotations.requestServerSyncStatus(forAccount: userAccount) { enableSync in
+      if enableSync {
+        libraryDetails.syncPermissionGranted = true
+      }
+
+      NYPLMainThreadRun.sync {
+        postWork(enableSync)
+      }
+
+      self.permissionsCheckLock.unlock()
+    }
+  }
+}
